Refactored StringOutput to use std::to_string Fixed Reference/Pointer bugs in all formulas. Implemented EigenDtmcPrctlModelChecker
Replaced uses of int32 with 64bits

+++ src/models/Dtmc.h
@@ -102,7 +102,7 @@ public:
 	/*!
 	 *
 	 */
-	std::set<std::string> getPropositionsForState(uint_fast32_t state) {
+	std::set<std::string> getPropositionsForState(uint_fast64_t state) {
 		return stateLabeling->getPropositionsForState(state);
 	}
